refactor: (home)ディレクトリの中身をスネークケースでリネーム
