--- src/app/(home)/_components/HomeTemplate/index.ts
@@ -1,1 +1,0 @@
-export { HomeTemplate } from './HomeTemplate'
