--- src/app/(home)/_components/NewsListSlider/index.ts
@@ -1,1 +1,0 @@
-export { NewsListSlider } from './NewsListSlider'
