+++ src/app/(home)/_components/home-template/index.ts
@@ -1,0 +1,1 @@
+export { HomeTemplate } from './home-template'
