+++ src/app/(home)/_components/news-list-slider/index.ts
@@ -1,0 +1,1 @@
+export { NewsListSlider } from './news-list-slider'
