+++ src/app/(home)/_components/news-list-slider/news-list-slider.test.tsx
@@ -1,0 +1,10 @@
+// テストファイルはこういうふうに、対象ファイルの隣におくんだよってことを示すためだけに一旦つくってる
+// TODO: あとでちゃんとテスト書く
+
+import { describe, it } from "vitest";
+
+describe('ニュース一覧（スライダー）', () => {
+  it('仮テスト', () => {
+
+  })
+})
