+++ src/app/(home)/_components/news-list-slider/news-list-slider.tsx
